+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * UI tests replacement with RTL/Jest. Refs FAT-37.
+* Export combined Fund code list. Refs UIF-320.
 
 ## [2.4.2](https://github.com/folio-org/ui-finance/tree/v2.4.2) (2021-07-28)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v2.4.1...v2.4.2)